packages/python-app/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-cloud-aiplatform >= 1.31.0
google-cloud-logging
vertexai
deta